+++ src/datagen/java/blusunrize/immersiveengineering/data/recipes/IngredientRecipes.java
@@ -75,6 +75,7 @@ public class IngredientRecipes extends IERecipeProvider
 				.group("sticks")
 				.unlockedBy("has_netherite_ingot", has(Tags.Items.INGOTS_NETHERITE))
 				.save(out, toRL(toPath(Ingredients.STICK_NETHERITE)));
+		add3x3Conversion(Items.NETHERITE_INGOT, Ingredients.NUGGET_NETHERITE, IETags.netheriteNugget, out);
 		shapedMisc(Ingredients.HEMP_FABRIC)
 				.pattern("fff")
 				.pattern("fsf")